Fix hunyuan_video_1_5_i2v_720p_gguf: use float timestep for V1.5 transformer
HunyuanVideo15TimestepEmbedding.forward casts its projections with
.to(dtype=timestep.dtype), so a long timestep propagates into bfloat16
linear layers and causes a dtype mismatch. Use a float timestep as the
V1.0 loaders do.

+++ hunyuan_video_1_5_i2v_720p_gguf/pytorch/loader.py
@@ -154,7 +154,7 @@ class ModelLoader(ForgeModel):
             config.image_embed_dim,
             dtype=dtype,
         )
-        timestep = torch.tensor([500], dtype=torch.long).expand(batch_size)
+        timestep = torch.tensor([500.0], dtype=dtype).expand(batch_size)
 
         return {
             "hidden_states": hidden_states,
